requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
distro==1.9.0
fastapi==0.129.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==1.2.13
langchain-openai==1.1.10
langsmith==0.7.4
lxml==6.0.2
openai==2.21.0
orjson==3.11.7
packaging==26.0
pillow==12.1.1
pydantic==2.12.5
pydantic_core==2.41.5
PyMuPDF==1.27.1
python-dotenv==1.2.1
python-multipart==0.0.22
python-pptx==1.0.2
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
sniffio==1.3.1
starlette==0.52.1
tenacity==9.1.4
tiktoken==0.12.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.0
uvicorn==0.41.0
xlsxwriter==3.2.9
xxhash==3.6.0
zstandard==0.25.0